Updates for Snap support
The LXD binary has been replaced with a script telling you to install
the lxd snap in newer versions of Ubuntu. We need to handle the snap
version of lxd better.

These changes check for the LXD snap and work with it in a much more
comprehensive way. For example, we use 'snap restart' to manage the
service. I also had to add a ridiculous amount of checking for the
container's state since lxc may or may not return an error code if it
fails to start a container.