crossbuilder: fix missing moc tool for cmake
It's not clear how cmake finds the `moc` tool, but for some reason it
expects to find it in `/usr/lib/$HOST_FARCH/qt5/bin/`. So, let's put it
there.

Also, rework a bit the way that qmake is setup, by using a more robust
approach that does not require the `QT_SELECT=qt5` trick: we simply
install qmake in `/usr/local/bin/`, just like `clickable` is doing.

Last but not least, correct the value of the `HOST_FARCH` variable,
which was computed wrong (it is used in this part of crossbuilder
anyway, so changing it should not bring any side-effects).