Bump version with +1 in front (#76)
Due to the way our CI works, the version in the changelog will usually
be bumped with `+0`. So, to combat with that, bump ours to `+1`.

This is particularly important when working with a library. When other
packages are built against this one, dpkg-shlibdeps embeds the version
of this package as seen in the CI. If we don't get this package's
version to be higher, apt will start removing those packages installed
on the phone.